requirements.txt
Django==3.0.8
psycopg2==2.8.5
coverage==5.2.1
python-dateutil==2.8.1
